reduceh is working!

+++ libvips/include/vips/resample.h
@@ -46,8 +46,8 @@ int vips_shrinkv( VipsImage *in, VipsImage **out, int yshrink, ... );
 
 int vips_reduce( VipsImage *in, VipsImage **out, 
 	double xshrink, double yshrink, ... );
-int vips_reduceh( VipsImage *in, VipsImage **out, int xshrink, ... );
-int vips_reducev( VipsImage *in, VipsImage **out, int yshrink, ... );
+int vips_reduceh( VipsImage *in, VipsImage **out, double xshrink, ... );
+int vips_reducev( VipsImage *in, VipsImage **out, double yshrink, ... );
 
 int vips_similarity( VipsImage *in, VipsImage **out, ... )
 	__attribute__((sentinel));
